+++ src/components/Button/index.css
@@ -1,0 +1,11 @@
+.button {
+    height: 35px;
+    width: 100px;
+    border: none;
+    background-color: #109FEC;
+    color: white;
+    border-radius: 50px;
+    font-size: 14px;
+    padding-left: 0px;
+    padding-right: 0px;
+}
