--- src/components/FormSubmitButton/index.css
@@ -1,12 +1,0 @@
-.footer_button {
-    height: 35px;
-    width: 100px;
-    border: none;
-    background-color: #109FEC;
-    color: white;
-    border-radius: 50px;
-    font-size: 14px;
-    position: fixed;
-    bottom: 20px;
-    right: 25px;
-}
