+++ src/containers/Authentication/PasswordField/index.css
@@ -7,7 +7,6 @@
 .label {
     font-size: 14px;
     color: grey;
-    padding-left: 28px;
     text-align: left;
     padding-bottom: 15px;
 }
