+++ src/containers/Authentication/SubmitButton/index.css
@@ -1,0 +1,5 @@
+.floating-button {
+    position: fixed;
+    bottom: 20px;
+    right: 25px;
+}
