+++ src/containers/Authentication/SubmitButton/index.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import * as PropTypes from 'prop-types';
+import './index.css';
+import Button from '../../../components/Button';
+
+const SubmitButton = (props) => {
+    return (
+        <div className="floating-button">
+            <Button
+                button_name="LOGIN"
+                onClick={props.onClick}/>
+        </div>
+    );
+};
+
+SubmitButton.propTypes = {
+    onClick: PropTypes.func.isRequired,
+};
+
+export default SubmitButton;
