+++ src/containers/Loader/index.js
@@ -2,9 +2,9 @@ import * as PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 import { load } from '../../actions/application';
-import AppAuthentication from '../Authentication';
+import Authentication from '../Authentication';
 
-class Application extends React.Component {
+class Loader extends React.Component {
     componentDidMount () {
         this.props.load();
     }
@@ -16,13 +16,13 @@ class Application extends React.Component {
 
         return (
             <div className="App">
-                <AppAuthentication/>
+                <Authentication/>
             </div>
         );
     }
 }
 
-Application.propTypes = {
+Loader.propTypes = {
     isLoading: PropTypes.bool.isRequired,
     load: PropTypes.func.isRequired,
 };
@@ -37,4 +37,4 @@ const actionsToProps = {
     load,
 };
 
-export default connect(stateToProps, actionsToProps)(Application);
+export default connect(stateToProps, actionsToProps)(Loader);
